lz4: guard Go-1.10 specific API behind build tags
Fixes #24.
